// frontend/src/metabase/query_builder/components/Canvas/types.ts
export type LanguageOptions =
  | "english"
  | "mandarin"
  | "spanish"
  | "french"
  | "hindi";

export type ProgrammingLanguageOptions =
  | "typescript"
  | "javascript"
  | "cpp"
  | "java"
  | "php"
  | "python"
  | "html"
  | "sql"
  | "other";
// ...
export type ArtifactLengthOptions = "shortest" | "short" | "long" | "longest";

export type ReadingLevelOptions =
  | "pirate"
  | "child"
  | "teenager"
  | "college"
  | "phd";
// ...
export interface ArtifactV3 {
  currentIndex: number;
  contents: (ArtifactMarkdownV3 | ArtifactCodeV3 | ArtifactQueryV3 | ArtifactInsightsV3)[];
}

export interface ArtifactMarkdownV3 {
  index: number;
  type: "text";
  title: string;
  description?: string;
  fullMarkdown: string;
}

export interface ArtifactCodeV3 {
  index: number;
  type: "code";
  title: string;
  description?: string;
  language: ProgrammingLanguageOptions;
  code: string;
}

export interface ArtifactQueryV3 {
  index: number;
  type: "card";
  title: string;
  cardId: any;
  queryResult?: any;
  rawResponse?: any;
}

export interface ArtifactImageV3 {
  index: number;
  type: "image";
  title: string;
  image: any;
}

export interface ArtifactInsightsV3 {
  index: number;
  type: "insight";
  title: string;
  insight: any;
}

export interface CodeHighlight {
  startCharIndex: number;
  endCharIndex: number;
}

export interface TextHighlight {
  fullMarkdown: string;
  markdownBlock: string;
  selectedText: string;
}